Move import action from landing screen to toolbar
