+++ packages/devtools_app/lib/src/framework/scaffold.dart
@@ -60,6 +60,7 @@ class DevToolsScaffold extends StatefulWidget {
   static List<Widget> defaultActions() => const [
         OpenSettingsAction(),
         ReportFeedbackButton(),
+        ImportToolbarAction(),
         OpenAboutAction(),
       ];
